feat(codec): MT filtered members + fix MT long-range self-shadow
Filter members (auto delta/x86) encoded the whole member sequentially
even with threads available — the 13 MB ntoskrnl DLL took 6.7 s at m3
where the unfiltered path used every core. Now:

- encode_with_filters_mt runs the forward transform (unchanged) then the
  windowed MT encode over the transformed member, with the filter
  records leading the first slice's symbol stream (member-relative
  positions preserved). encode_with_auto_delta_filter and
  encode_with_auto_x86_filter take a threads count and use it for the
  full member encodes; the sample-based variant selection stays
  sequential (64 KiB).
- The MT path's long-range table is pre-built over the whole window, so
  a slice's own positions shadow the copy source it should match — a
  32 MiB file of random data plus an exact distant copy compressed to
  ~100% instead of 50%. The table now keeps the previous occurrence per
  key (vals2) and probes with get_before, so a self-shadowing probe
  falls back to the real source. The sequential encoder never consults
  the fallback (its newest entry is always before the chunk being
  parsed), so its output is byte-identical (verified).

Measured (m3): dll 6.7 -> 3.1 s at -mt8 (and 36 KB smaller), mixed
2.7 -> 1.1 s, text 0.73 -> 0.39 s; pair32 distant-copy 33.5 -> 16.8 MB.
All 120 lib tests + CLI suite green; decode roundtrips byte-identical.

+++ crates/rar/src/codec/filters.rs
@@ -703,7 +703,7 @@ mod tests {
                 state ^= state >> 27;
                 let r = (state >> 33) as u32;
                 *v += (r % 8) as i64 - 4;
-                let value = *v as i64;
+                let value = *v;
                 for b in 0..bytes {
                     out.push((value >> (8 * b)) as u8);
                 }
